requirements.txt
vanna[chromadb,openai,postgres]
pandas
python-dotenv